create Controller, Repo and Service for user. Rename UserData.java class to User.java
