+++ backend/src/main/java/org/example/backend/model/User.java
@@ -1,21 +1,29 @@
 package org.example.backend.model;
 
+import lombok.Data;
+
+@Data
 public class User {
+    private String _id;
     private String firstname;
     private String lastname;
     private String email;
     private String password;
-    private double weight;
-    private double height;
+    private double weightInKg;
+    private double heightInCm;
     private double caloriesEatPerDay;
     private double targetWeightReduce;
     private double targetTimeInWeek;
-    private enum activity
-    {LOW (0.2) , MIDDLE (0.4), HIGH (0.6);
-        private final double value; activity (double value){this.value=value;}}
-    private Workout[] WorkoutPlan;
-    private double Bmi = weight / ((height/100) * (height/100));
-    private double caloriesUsedPerDayFromActivityAndWeight = (weight * 24 /* +activity.g*/);
+    private enum activity {
+        LOW (0.2) , MIDDLE (0.4), HIGH (0.6);
+        private final double value; activity (double value){this.value=value;}
+        public double getValue() {
+            return value;
+        }
+    }
+    private Workout[] workoutPlan;
+    private double bmi = weightInKg / ((heightInCm /100) * (heightInCm /100));
+    private double caloriesUsedPerDayFromActivityAndWeight = (weightInKg * 24 /* +activity.g*/);
     private double caloriesOverflowPerDay = caloriesEatPerDay - caloriesUsedPerDayFromActivityAndWeight;
     private double caloriesNeedToReducePerWeekForTargetWeightReduce = (caloriesOverflowPerDay * 7) + ((targetWeightReduce * 7000) / targetTimeInWeek);
 
@@ -24,90 +32,11 @@ public class User {
                 double targetWeightReduce, double targetTimeInWeek){
         this.firstname = firstname;
         this.lastname = lastname;
-        this.weight = weight;
-        this.height = height;
+        this.weightInKg = weight;
+        this.heightInCm = height;
         this.caloriesEatPerDay = caloriesEatPerDay;
         this.targetWeightReduce = targetWeightReduce;
-        this.targetTimeInWeek = targetTimeInWeek;    }
-
-
-    public String getFirstname() {
-        return firstname;
-    }
-
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
-    }
-
-    public String getLastname() {
-        return lastname;
-    }
-
-    public void setLastname(String lastname) {
-        this.lastname = lastname;
-    }
-
-    public double getWeight() {
-        return weight;
-    }
-
-    public void setWeight(double weight) {
-        this.weight = weight;
-    }
-
-    public double getHeight() {
-        return height;
-    }
-
-    public void setHeight(double height) {
-        this.height = height;
-    }
-
-    public double getCaloriesEatPerDay() {
-        return caloriesEatPerDay;
-    }
-
-    public void setCaloriesEatPerDay(double caloriesEatPerDay) {
-        this.caloriesEatPerDay = caloriesEatPerDay;
-    }
-
-    public double getTargetWeightReduce() {
-        return targetWeightReduce;
-    }
-
-    public void setTargetWeightReduce(double targetWeightReduce) {
-        this.targetWeightReduce = targetWeightReduce;
-    }
-
-    public double getTargetTimeInWeek() {
-        return targetTimeInWeek;
-    }
-
-    public void setTargetTimeInWeek(double targetTimeInWeek) {
         this.targetTimeInWeek = targetTimeInWeek;
     }
 
-    public Workout[] getWorkoutPlan() {
-        return WorkoutPlan;
-    }
-
-    public void setWorkoutPlan(Workout[] workoutPlan) {
-        WorkoutPlan = workoutPlan;
-    }
-
-    public double getBmi() {
-        return Bmi;
-    }
-
-    public double getCaloriesUsedPerDayFromActivityAndWeight() {
-        return caloriesUsedPerDayFromActivityAndWeight;
-    }
-
-    public double getCaloriesOverflowPerDay() {
-        return caloriesOverflowPerDay;
-    }
-
-    public double getCaloriesNeedToReducePerWeekForTargetWeightReduce() {
-        return caloriesNeedToReducePerWeekForTargetWeightReduce;
-    }
 }
