+++ backend/src/main/java/org/example/backend/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package org.example.backend.repository;
 
+import org.example.backend.model.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
