+++ backend/src/main/java/org/example/backend/service/UserService.java
@@ -1,24 +1,23 @@
 package org.example.backend.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.backend.model.User;
 import org.example.backend.repository.UserRepository;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
 private final UserRepository repo;
 
-    public List<User> getUserById(){
-        return repo.findAll();
+    public User getUserById(String id){
+        return repo.findById(id).orElseThrow();
     }
 
-    public User saveNewUser(User user) {
+/*    public User saveNewUser(User user) {
         repo.save(user);
         return repo.findById(user.getId()).orElseThrow();
-    }
+    }*/
 
     /*public User updateUserById(User user) {
         repo.save(user);
